+++ src/main/java/team18/team18_be/auth/dto/request/MemberIdRequest.java
@@ -1,4 +1,4 @@
-package team18.team18_be.auth.dto;
+package team18.team18_be.auth.dto.request;
 
 public record MemberIdRequest(
     Long id
