+++ src/main/java/team18/team18_be/auth/dto/request/UserTypeRequest.java
@@ -1,4 +1,4 @@
-package team18.team18_be.auth.dto;
+package team18.team18_be.auth.dto.request;
 
 public record UserTypeRequest(
     int type
