+++ src/main/java/team18/team18_be/auth/dto/response/OAuthJwtResponse.java
@@ -1,4 +1,4 @@
-package team18.team18_be.auth.dto;
+package team18.team18_be.auth.dto.response;
 
 public record OAuthJwtResponse(
     String accessToken,
